+++ src/models/targets.v
@@ -73,4 +73,6 @@ pub mut:
 	limit  u64 = 25
 	offset u64
 	repo   string
+	query  string
+	arch   string
 }